Fix #19: Improve command path search
The in-process execution assumed it could find a module's scripts in the
same location as the Python interpreter. However, when installing a
module system-wide, many distros install the scripts to a different path
than the interpreter, e.g. `/usr/local/bin` instead of `/usr/bin`.

Instead, use `distutils.spawn.find_executable` to find the path to the
requested script. In addition to searching the system path, this method
can also find executables relative to the working directory (in this
case, the path from which pytest is invoked).

`distutils.spawn.find_executable` is present in Python >=2.6 and >=3.0.