Trigger tests on push and on pull request activity.
I like to use the CI to test my topic branches before making a
PR for them, but I can't do that when push is restricted to the master
branch.

Having push and pull_request can cause updates to PR's to be tested
twice when updated.  Restricting the types reduces this.

To manually trigger a test on a PR you can close and reopen that PR.