Move all setup metadata to pyproject.toml
I used Flit as the build tool, this removes `setup.py` entirely.
This changes versioning slightly as the version has to be at `__version__` in the module.
I'm not sure if this is avoidable, but it's pretty standard to do it this way.

I also added a few relevant project URL's.
Such as the issues page and the changelog.
These will show up on PyPI for the next release after this is merged.

Remove references to setup.py from tox.